% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAUCscore_function.R
\name{scAUCscore}
\alias{scAUCscore}
\title{Calculate CD8 TIL states enrichment scores (required to run predictTilState function)}
\usage{
scAUCscore(sce, nCores = 1, sigs, aucMaxRank = 1500)
}
\arguments{
\item{nCores}{number of cores to use in parallel calculation of gene rankings}

\item{data}{single-cell expression matrix. Only gene expression ranks in each cell will be used and therefore any cell-to-cell normalization method used is not relevant (e.g. UMI counts, CPM, TPM, TMM)}
}
\description{
\code{scAUCscore} This function evaluates a logistic regression model to predict the state of individual CD8 tumor-infiltrating lymphocytes (mouse or human) based on their transcriptomes (scRNA-seq data)
}
